Fix bad input fd error
This fixes the following error:

22:00:08.129 [error] Bad input fd in erts_poll()! fd=0, port=#Port<0.639>, driver=tty_sl, name=tty_sl -c -e

It also removes the information from the README.md. That information
turned out to be incorrect. It was possible to get the "Bad input fd"
error pretty much anywhere given enough patience. Based on the fix,
single core machines were immune and anecdotally that was actually
observed.